move tool utils to common
