+++ common/build.gradle
@@ -15,6 +15,7 @@ plugins {
 }
 
 dependencies {
+    implementation project(path: ":${rootProject.name}-spi", configuration: 'shadow')
     compileOnly group: 'org.opensearch', name: 'opensearch', version: "${opensearch_version}"
     compileOnly group: 'org.reflections', name: 'reflections', version: '0.9.12'
     testImplementation group: 'junit', name: 'junit', version: '4.13.2'
